c/clang/symg:move get conf func
